export 'local_props.dart';
